backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
sqlalchemy==2.0.40
pymysql==1.1.1
alembic==1.15.2
pydantic-settings==2.8.1
email-validator==2.2.0
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
celery[redis]==5.5.1
redis==5.2.1
python-multipart==0.0.20
httpx==0.28.1
openai==1.107.2
pytest==8.3.5
ruff==0.11.13
